Implement autobot testing for set replication (#336)
* Implement autobot testing for set replication

* Address comments

* Fix style check

* Address comments to add waitFor rather than a hard-coded timeout

* Address comments

* Address comment

+++ job/microbench/src/main/java/alluxio/job/benchmark/AsyncWriteDefinition.java
@@ -160,7 +160,7 @@ public final class AsyncWriteDefinition
           throw Throwables.propagate(e);
         }
       }
-    }, config.getPersistTimeout() * Constants.SECOND_MS);
+    }, (long) config.getPersistTimeout() * Constants.SECOND_MS);
   }
 
   @Override
